disp_training/requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
ansi2html==1.8.0
antlr4-python3-runtime==4.9.3
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.3
attrs==23.1.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4==4.12.2
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click @ file:///croot/click_1698129812380/work
cloudpickle @ file:///croot/cloudpickle_1683040006038/work
colorama @ file:///croot/colorama_1672386526460/work
comm==0.1.4
ConfigArgParse==1.7
contourpy @ file:///croot/contourpy_1700583582875/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
cytoolz @ file:///croot/cytoolz_1701723583781/work
dash==2.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask @ file:///croot/dask-core_1701396095060/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
einops==0.7.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastjsonschema==2.18.0
filelock==3.12.2
Flask==2.2.5
fonttools==4.25.0
frozenlist==1.4.1
fsspec @ file:///croot/fsspec_1701286474621/work
gdown==4.7.1
grpcio==1.60.1
h5py @ file:///croot/h5py_1691589708553/work
hdf5plugin @ file:///home/conda/feedstock_root/build_artifacts/hdf5plugin_1651556479333/work
hydra-core==1.3.2
idna==3.6
imagecodecs @ file:///croot/imagecodecs_1695064943445/work
imageio @ file:///croot/imageio_1707247282708/work
importlib-metadata==6.8.0
importlib-resources @ file:///croot/importlib_resources-suite_1704281845041/work
ipython @ file:///croot/ipython_1674681422581/work
ipywidgets==8.1.0
itsdangerous==2.1.2
jedi @ file:///tmp/build/80754af9/jedi_1644297102865/work
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
lightning==2.2.0.post0
lightning-utilities==0.10.1
locket @ file:///opt/conda/conda-bld/locket_1652903118915/work
Markdown==3.5.2
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib @ file:///croot/matplotlib-suite_1679593461707/work
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mpmath==1.3.0
multidict==6.0.5
munkres==1.1.4
nbformat==5.7.0
nest-asyncio==1.5.7
networkx @ file:///croot/networkx_1690561992265/work
numpy==1.25.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open3d==0.17.0
packaging @ file:///croot/packaging_1693575174725/work
pandas==2.1.0
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
partd @ file:///croot/partd_1698702562572/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
pillow @ file:///croot/pillow_1707233021655/work
platformdirs==3.10.0
plotly==5.16.1
ply==3.11
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
protobuf==4.25.3
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
Pygments @ file:///croot/pygments_1684279966437/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5==5.15.10
PyQt5-sip @ file:///croot/pyqt-split_1698769088074/work/pyqt_sip
pyquaternion==0.9.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytorch-lightning==2.2.0.post0
pytz==2023.3
pywavelets @ file:///croot/pywavelets_1705049820073/work
PyYAML @ file:///croot/pyyaml_1698096049011/work
referencing==0.30.2
requests==2.31.0
retrying==1.3.4
rpds-py==0.10.0
SciencePlots==2.1.1
scikit-image @ file:///croot/scikit-image_1669241743693/work
scikit-learn==1.3.0
scipy==1.11.2
sip @ file:///croot/sip_1698675935381/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
soupsieve==2.4.1
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tifffile @ file:///croot/tifffile_1695107451082/work
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
toolz @ file:///croot/toolz_1667464077321/work
torch==2.2.0
torchmetrics==1.3.1
torchvision==0.17.0
tornado @ file:///croot/tornado_1696936946304/work
tqdm==4.66.1
traitlets @ file:///croot/traitlets_1671143879854/work
triton==2.2.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.4
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
Werkzeug==2.2.3
widgetsnbextension==4.0.8
yarl==1.9.4
zipp @ file:///croot/zipp_1704206909481/work
